Merch inv show

+++ db/migrate/20210925202746_create_bulk_discounts.rb
@@ -2,7 +2,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
       t.integer :quantity
-      t.decimal :percentage
+      t.integer :percentage
       t.references :merchant, foreign_key: true
 
       t.timestamps
